restructure app-ui code
