+++ app-ui/src/components/DeploymentInfo.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Modal } from "../Modal";
+import { Modal } from "../lib/Modal";
 import { config } from "../utils/Config";
 import { BsLockFill } from "react-icons/bs";
