+++ app-ui/src/pages/traces/AnnotationAugmentedTraceView.jsx
@@ -4,7 +4,10 @@ import { Tooltip } from "react-tooltip";
 import "./Annotations.scss";
 import "../../styles/Explorer.scss";
 
-import { RemoteResource, useRemoteResource } from "../../RemoteResource";
+import {
+  RemoteResource,
+  useRemoteResource,
+} from "../../service/RemoteResource";
 import { useUserInfo } from "../../utils/UserInfo";
 import UserIcon from "../../lib/UserIcon";
 
@@ -15,7 +18,7 @@ import { openInPlayground } from "../../lib/playground";
 import { HighlightedJSON } from "../../lib/traceview/highlights";
 import { RenderedTrace } from "../../lib/traceview/traceview";
 import { config } from "../../utils/Config";
-import { useTelemetry } from "../../telemetry";
+import { useTelemetry } from "../../utils/telemetry";
 import { AnnotationsParser } from "../../lib/annotations_parser";
 import { HighlightDetails } from "./HighlightDetails";
