+++ app-ui/src/pages/traces/Dataset.tsx
@@ -13,13 +13,16 @@ import {
 } from "react-icons/bs";
 import { Link, useLoaderData, useNavigate } from "react-router-dom";
 import { DeleteDatasetModalContent } from "../home/NewDataset";
-import { Modal } from "../../Modal";
+import { Modal } from "../../lib/Modal";
 import { PoliciesView } from "./Policies";
-import { RemoteResource, useRemoteResource } from "../../RemoteResource";
+import {
+  RemoteResource,
+  useRemoteResource,
+} from "../../service/RemoteResource";
 import { useUserInfo } from "../../utils/UserInfo";
 import { Metadata } from "../../lib/metadata";
 import { config } from "../../utils/Config";
-import { useTelemetry } from "../../telemetry";
+import { useTelemetry } from "../../utils/telemetry";
 import { DatasetNotFound, isClientError } from "../notfound/NotFound";
 import { Traces } from "./Traces";
