+++ app-ui/src/utils/telemetry.ts
@@ -1,7 +1,7 @@
 import posthog from "posthog-js";
-import { Plugins } from "./lib/traceview/plugins";
+import { Plugins } from "../lib/traceview/plugins";
 import { usePostHog } from "posthog-js/react";
-import { config } from "./utils/Config";
+import { config } from "./Config";
 
 export const SUPPORTS_TELEMETRY = config("telemetry");
 export const HAS_CONSENT = window.localStorage.getItem("consent") === "true";
